+++ httpd/auth/basic.go
@@ -57,7 +57,7 @@ func (b *Basic) Preauthenticate() (*http.Cookie, error) {
 		Name:     LoginCookie,
 		Value:    token,
 		Path:     "/",
-		MaxAge:   300 * int(time.Hour.Seconds()), // 5 minutes
+		MaxAge:   int((5 * time.Minute).Seconds()),
 		HttpOnly: true,
 		SameSite: http.SameSiteStrictMode,
 		//	Secure:   true,