+++ auth/auth.go
@@ -42,6 +42,7 @@ type IUser interface {
 	Password() ([]byte, string)
 	Role() string
 	Locked() bool
+	Login(err error)
 }
 
 type Authorizator struct {