+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+## [Unreleased]
+
 ## [0.8.2](https://github.com/uhppoted/uhppoted-httpd/releases/tag/v0.8.2) - 2022-10-14
 
 ### Changed