+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 2. Locked user login after too many failed attempts
 3. Removed legacy support for users in auth.json
 4. Updated _systemd_ unit file to wait for `network-online.target`
+5. Reworked lockfile to use `flock` _syscall_.
 
 ## [0.8.2](https://github.com/uhppoted/uhppoted-httpd/releases/tag/v0.8.2) - 2022-10-14