+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 3. Removed legacy support for users in auth.json
 4. Updated _systemd_ unit file to wait for `network-online.target`
 5. Reworked lockfile to use `flock` _syscall_.
+6. Moved default lockfile to system _temp_ folder.
 
 ## [0.8.2](https://github.com/uhppoted/uhppoted-httpd/releases/tag/v0.8.2) - 2022-10-14